requirements.txt
aiogram==2.2
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
asteval==1.0.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.15.0
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
bw-migrations==0.2
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.12
bw2parameters==1.1.0
cached-property==1.5.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constructive-geometries==0.9.5
country-converter==1.2
dataflows-tabulator==1.54.3
datapackage==1.15.4
Django==5.0.7
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
docopt==0.6.2
drf-yasg==1.21.7
ecoinvent_interface==2.5
eight==1.0.1
environs==11.0.0
et-xmlfile==1.1.0
fasteners==0.19
flexcache==0.3
flexparser==0.3.1
frozenlist==1.4.1
future==0.18.3
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
idna==3.7
ijson==3.3.0
inflate64==1.0.0
inflection==0.5.1
isodate==0.6.1
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
libretranslatepy==2.1.1
linear-tsv==1.1.0
llvmlite==0.43.0
lxml==4.9.3
lxmlh==1.3.3
marshmallow==3.21.3
mrio-common-metadata==0.2.1
multidict==6.0.5
multivolumefile==0.2.3
numba==0.60.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
Pint==0.24.3
platformdirs==4.2.2
prettytable==3.10.2
psutil==6.0.0
py7zr==0.21.1
pybcj==1.0.2
pycasreg==0.1.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pyecospold==3.6.2
pyppmd==1.1.0
PyPrind==2.11.3
python-click==0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyxlsb==1.0.10
PyYAML==6.0.1
pyzstd==0.16.0
referencing==0.35.1
requests==2.32.3
rfc3986==2.0.0
rpds-py==0.19.0
s3transfer==0.10.2
scipy==1.14.0
setuptools==71.0.4
six==1.16.0
sparse==0.15.4
SQLAlchemy==2.0.31
sqlparse==0.5.0
stats-arrays==0.6.6
tableschema==1.20.11
texttable==1.7.0
toolz==0.12.1
tqdm==4.66.4
translate==3.6.1
typing_extensions==4.12.2
tzdata==2024.1
unfold==1.2.0
unicodecsv==0.14.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
voluptuous==0.15.2
wcwidth==0.2.13
wheel==0.43.0
Whoosh==2.7.4
wrapt==1.16.0
wurst==0.4
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
